+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/repository/KeysSynchronization.kt
@@ -1,8 +1,12 @@
 package com.flipperdevices.bridge.synchronization.impl.repository
 
+import com.flipperdevices.bridge.dao.api.delegates.key.SimpleKeyApi
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyType
 import com.flipperdevices.bridge.synchronization.impl.di.TaskGraph
+import com.flipperdevices.bridge.synchronization.impl.model.DiffSource
+import com.flipperdevices.bridge.synchronization.impl.repository.android.AndroidHashRepository
 import com.flipperdevices.bridge.synchronization.impl.repository.flipper.TimestampSynchronizationChecker
+import com.flipperdevices.bridge.synchronization.impl.repository.manifest.ManifestRepository
 import com.flipperdevices.bridge.synchronization.impl.repository.manifest.ManifestTimestampRepository
 import com.flipperdevices.bridge.synchronization.impl.utils.ProgressWrapperTracker
 import com.flipperdevices.core.log.LogTagProvider
@@ -21,7 +25,10 @@ private const val PERCENT_FETCH_TIMESTAMP = 0.1f
 class KeysSynchronizationImpl @Inject constructor(
     private val folderKeySynchronization: FolderKeySynchronization,
     private val timestampSynchronizationChecker: TimestampSynchronizationChecker,
-    private val manifestTimestampRepository: ManifestTimestampRepository
+    private val manifestTimestampRepository: ManifestTimestampRepository,
+    private val manifestRepository: ManifestRepository,
+    private val androidHashRepository: AndroidHashRepository,
+    private val simpleKeyApi: SimpleKeyApi
 ) : KeysSynchronization, LogTagProvider {
     override val TAG = "KeysSynchronization"
 
@@ -38,7 +45,7 @@ class KeysSynchronizationImpl @Inject constructor(
         )
 
         val typesToUpdates = typesUpdateTimestamps.filter { (type, timestamp) ->
-            timestamp == null || manifestTimestampRepository.isUpdateRequired(type, timestamp)
+            isNeedUpdate(type, timestamp)
         }
         val percentForOneType = (1.0f - PERCENT_FETCH_TIMESTAMP) / typesToUpdates.size
         var currentMinPercent = PERCENT_FETCH_TIMESTAMP
@@ -56,5 +63,29 @@ class KeysSynchronizationImpl @Inject constructor(
                 manifestTimestampRepository.setTimestampForType(type, timestamp)
             }
         }
+    }
+
+    private suspend fun isNeedUpdate(type: FlipperKeyType, lastFolderUpdate: Long?): Boolean {
+        val androidKeys = simpleKeyApi.getExistKeys(type)
+        val hashes = androidHashRepository.getHashes(androidKeys)
+        val changedKeysOnAndroid = manifestRepository.compareFolderKeysWithManifest(
+            folder = type.flipperDir,
+            keys = hashes,
+            diffSource = DiffSource.ANDROID
+        )
+
+        if (changedKeysOnAndroid.isNotEmpty()) {
+            return true
+        }
+
+        if (lastFolderUpdate == null) {
+            return true
+        }
+
+        if (manifestTimestampRepository.isUpdateRequired(type, lastFolderUpdate)) {
+            return true
+        }
+
+        return false
     }
 }