+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [FIX] On `save dump as` save only editable key
 - [FIX] Remove emulate button on deleted screen
 - [FIX] Fix shadow file synchronization
+- [FIX] Fix broken synchronization with changes from android
 - [FIX] Fix shadow file sharing as file
 - [REFACTOR] Bump Android Gradle Plugin
 - [REFACTOR] Fix detekt compose issues